requirements.txt
bitsandbytes==0.43.1
jury==2.3.1
numpy==1.26.4
pandas==2.2.2
peft==0.10.0
Pillow==10.4.0
scikit_learn==1.3.0
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.1
torchvision==0.16.1
tqdm==4.65.0
transformers==4.38.2
trl==0.9.6
radgraph==0.1.2
